requirements.txt
numpy == 1.15.0
opencv-contrib-python == 3.4.5.20
wheel == 0.32.3
scikit-image == 0.14.1
pillow == 5.4.0
scipy == 1.2.0
h5py == 2.9.0
matplotlib == 3.0.2
blessings == 1.7
progressbar2 == 3.39.2
trimesh == 3.2.23
torch == 1.1.0
torchvision == 0.3.0
tensorboard == 1.14.0